django_backend/requirements.txt
aniso8601==9.0.1
asgiref==3.5.2
beautifulsoup4==4.10.0
bleach==5.0.1
blinker==1.4
boto3==1.26.104
botocore==1.29.104
Brlapi==0.8.3
cairocffi==1.4.0
CairoSVG==2.5.2
certifi==2020.6.20
cffi==1.15.1
chardet==4.0.0
click==8.0.3
colorama==0.4.4
command-not-found==0.3
cryptography==3.4.8
cssselect2==0.7.0
cupshelpers==1.0
dbus-python==1.2.18
defer==1.0.6
defusedxml==0.7.1
distlib==0.3.4
distro==1.7.0
distro-info===1.1build1
Django==4.1.4
django-cors-headers==4.0.0
django-storages==1.13.2
djangorestframework==3.14.0
djongo==1.3.6
dnspython==2.3.0
et-xmlfile==1.1.0
filelock==3.6.0
flake8==6.0.0
gdown==4.6.0
graphene==3.2.1
graphene-django==3.0.0
graphql-core==3.2.3
graphql-relay==3.2.0
gunicorn==20.1.0
html5lib==1.1
httplib2==0.20.2
idna==3.3
importlib-metadata==4.6.4
jeepney==0.7.1
jmespath==1.0.1
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
louis==3.20.0
lxml==4.8.0
macaroonbakery==1.3.1
Markdown==3.3.4
martor==1.6.15
mccabe==0.7.0
meld==3.20.4
more-itertools==8.10.0
netifaces==0.11.0
numpy==1.21.5
oauthlib==3.2.0
olefile==0.46
openpyxl==3.0.10
pandas==1.5.1
pexpect==4.8.0
Pillow==9.0.1
pipenv==11.9.0
platformdirs==2.5.1
promise==2.3
protobuf==3.12.4
ptvsd==4.3.2
ptyprocess==0.7.0
pycairo==1.20.1
pycodestyle==2.10.0
pycparser==2.21
pycups==2.0.1
pyflakes==3.0.1
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
pymongo==3.12.3
PyNaCl==1.5.0
pyparsing==2.4.7
pypng==0.20220715.0
PyQRCode==1.2.1
pyRFC3339==1.1
PySocks==1.7.1
python-apt==2.4.0+ubuntu1
python-dateutil==2.8.1
python-debian===0.1.43ubuntu1
python-dotenv==1.0.0
python-snappy==0.6.1
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
reportlab==3.6.8
requests==2.25.1
s3transfer==0.6.0
scour==0.38.2
SecretStorage==3.3.1
six==1.16.0
soupsieve==2.3.1
sqlparse==0.2.4
ssh-import-id==5.11
systemd-python==234
text-unidecode==1.3
tinycss2==1.2.1
tqdm==4.64.1
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.5
virtualenv==20.13.0+ds
virtualenv-clone==0.3.0
wadllib==1.3.6
webencodings==0.5.1
xdg==5
xkit==0.0.0
zipp==1.0.0
